# docs/swagger.json
{
    "components": {
        "schemas": {
...
            "domain.User": {
                "properties": {
                    "createdAt": {
                        "type": "string"
                    },
                    "defaultCurrency": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "updatedAt": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "http.AddFriendRequest": {
                "properties": {
                    "friendEmail": {
                        "type": "string"
                    },
                    "friendPhone": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "http.ListFriendsResponse": {
                "properties": {
                    "data": {
                        "items": {
                            "$ref": "#/components/schemas/domain.User"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "pagination": {
                        "$ref": "#/components/schemas/pagination.Meta"
                    }
                },
                "type": "object"
            },
...
            "http.RegisterRequest": {
                "properties": {
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "http.UpdateProfileRequest": {
                "properties": {
                    "defaultCurrency": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "pagination.Meta": {
                "properties": {
                    "hasMore": {
                        "type": "boolean"
                    },
                    "nextCursor": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "response.ErrorResponse": {
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "scheme": "bearer",
                "type": "http"
            }
        }
    },
    "info": {
        "description": "Backend API service for Splittr bill splitting.",
        "title": "Splittr API",
        "version": "1.0"
    },
...
    "paths": {
...
        "/friends": {
            "get": {
                "description": "Get a cursor-paginated list of the current user's friends.",
                "parameters": [
                    {
                        "description": "Items per page (max 100, default 20)",
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Opaque cursor token from a previous response",
                        "in": "query",
                        "name": "cursor",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/http.ListFriendsResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "List friends",
                "tags": [
                    "friends"
                ]
            },
            "post": {
                "description": "Create a friendship link with another user by their email or phone.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/http.AddFriendRequest",
                                        "summary": "request",
                                        "description": "Friend email or phone"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Friend email or phone",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/domain.User"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Add friend",
                "tags": [
                    "friends"
                ]
            }
        },
        "/friends/{friendId}": {
            "delete": {
                "description": "Delete a friendship link by friend ID.",
                "parameters": [
                    {
                        "description": "Friend ID",
                        "in": "path",
                        "name": "friendId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Remove friend",
                "tags": [
                    "friends"
                ]
            }
        },
...
        "/users": {
            "post": {
                "description": "Create a new user profile using Firebase identities.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/http.RegisterRequest",
                                        "summary": "request",
                                        "description": "Registration data"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Registration data",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/domain.User"
                                }
                            }
                        },
                        "description": "Created"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Register user",
                "tags": [
                    "users"
                ]
            }
        },
        "/users/me": {
            "get": {
                "description": "Retrieve the profile details of the logged-in user.",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/domain.User"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Get current user profile",
                "tags": [
                    "users"
                ]
            },
            "put": {
                "description": "Update name or default currency for the current user.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/http.UpdateProfileRequest",
                                        "summary": "request",
                                        "description": "Profile details to update"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Profile details to update",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/domain.User"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Update user profile",
                "tags": [
                    "users"
                ]
            }
        }
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "/v1"
        }
    ]
}